# docs/thv-registry-api/swagger.yaml
components:
...
  securitySchemes:
    BearerAuth:
      description: 'OAuth 2.0 Bearer token authentication. Format: "Bearer {token}"'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    url: https://github.com/stacklok/toolhive
  description: |-
    API for accessing MCP server registry data and deployed server information
    This API provides endpoints to query the MCP (Model Context Protocol) server registry,
    get information about available servers, and check the status of deployed servers.

    Authentication is required by default. Use Bearer token authentication with a valid
    OAuth/OIDC access token. The /.well-known/oauth-protected-resource endpoint provides
    OAuth discovery metadata (RFC 9728).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ToolHive Registry API
  version: "0.1"
openapi: 3.1.0
paths:
...
  /extension/v0/registries:
    get:
      description: List all registries
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not implemented
      security:
      - BearerAuth: []
      summary: List registries
      tags:
      - extension
  /extension/v0/registries/{registryName}:
    delete:
      description: Delete a registry
      parameters:
      - description: Registry Name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not implemented
      security:
      - BearerAuth: []
      summary: Delete registry
      tags:
      - extension
    get:
      description: Get a registry by name
      parameters:
      - description: Registry Name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not implemented
      security:
      - BearerAuth: []
      summary: Get registry
      tags:
      - extension
    put:
      description: Create or update a registry
      parameters:
      - description: Registry Name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not implemented
      security:
      - BearerAuth: []
      summary: Create or update registry
      tags:
      - extension
  /extension/v0/registries/{registryName}/servers/{serverName}/versions/{version}:
    put:
      description: Create or update a server in the registry
      parameters:
      - description: Registry Name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      - description: URL-encoded server name (e.g., \
        in: path
        name: serverName
        required: true
        schema:
          type: string
      - description: URL-encoded version to retrieve (e.g., \
        in: path
        name: version
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not implemented
      security:
      - BearerAuth: []
      summary: Create or update server
      tags:
      - extension
...